Add risc-v cheri's Sealed bit
Add a field for the Sealed bit that is defined in the Cheri risc-v cheri
specification. Enable this field for 64r and 128r, disable it for all
other formats.

For 64r, the Sealed bit sits in the middle of a reserved area. We have to
split this area in two parts which we call reserved and reserved2.

Define the wrappers for the sealed bit. We have to use uint8_t instead of
bool to avoid compiler warnings such as

cheri_compressed_cap_common.h:263:32: warning: comparison of
   constant ‘1’ with boolean expression is always true [-Wbool-compare]
         _cc_debug_assert(value <= _CC_N(FIELD_##X##_MAX_VALUE)); \
                                ^~

FWIW we can still pass _cc_N(get_sealed)'s uint8_t return value to a
function that expects a bool. The implicit conversion is defined by the
C99 standard.