Add AP field
Add the AP (Architectural Permissions) field as defined in the Cheri
risc-v cheri specification.

Enable the AP field for the 64r and 128r formats which implement Cheri
risc-v cheri.

This patch adds only the field definitions for AP. AP's encoding differs
between the 64r and 128r versions. We'll add helpers for the different
encodings in a separate commit.