Remove otype field for 64r and 128r
Remove the "otype" field from the 64r and 128r capability formats. Cheri
risc-v cheri capabilities do not contain an otype field.

Add format-specific defines to indicate which formats use the otype
field. Users of cheri-compressed-cap can check the define for their
selected format.

Please note that if we enabled the 64r and 128r formats in qemu at this
point, the handling of sealed capabilities would be broken. We need
either the otype (for v9) or the sealed bit (for risc-v cheri). Their
locations overlap so we have to remove otype first and then add sealed
in a subsequent commit.