Fix up asserts about exponent length
Fix up asserts that check the length of some exponent fields. For cheri
risc-v cheri with MXLEN=32, both TE and BE are 2 bytes long (and the exponent
is calculated from T8, TE and BE).

While at it, move the asserts before the comment about exact bounds. This
comment refers to the "exact" variable, not to the asserts.