Update null and infinite caps for risc-v cheri
Update the definitions of the null capability and the infinite capability
for risc-v cheri.

We can include all v9 and risc-v cheri fields in the definitions. Unused
fields are 0 and will be ignored.

Revert the exponent, risc-v cheri stores the difference to the maximum
exponent instead of the exponent's value.

With all fields set up correctly, the null capability is 0x0 for the
risc-v cheri formats 64r and 128r.