Update compute_ebt for risc-v cheri
Update the _cc_N(compute_ebt) function template for risc-v cheri.

Set either EF or IE. Use the reverted format for risc-v cheri's exponent,
split it between T8, TE and BE. Add an assert to check if the exponent is
out of range, this should not happen with the existing code.