store lvbits for each decompressed capability
risc-v cheri v0.9.0 adds an optional Zcherilevels extension.

The cheri-compressed-cap library has to know if Zcherilevels is supported
by the application that uses the library and if so, how many levels are
used. This affects some permissions and capability fields.

cheri-compressed-cap has no global init function and no context that is
preserved between calls to library functions. The easiest way to store
info about Zcherilevels support is to add a new field to the decompressed
capability struct. cr_lvbits stores the number of supported levels or 0
if Zcherilevels is not used.

Callers that create or decompress a capability have a pass a value for
lvbits. This will be addressed in subsequent commits.