lvbits parameter for inifinite capability function
Add a new lvbits parameter to the inifinite capability function. Only the
caller knows if Zcherilevels are supported and if so, how many levels are
used.

Extend the existing function that takes basic capability fields and a
flag for Zcherihybrid support. Technically, this is breaking the
library's external API. However, the version with just the Zcherihybrid
flag hasn't been widely used and has not been supported in a public
release.

For now, we limit lvbits to 1. This gives us two levels (local, global).
The risc-v cheri spec does not cover other setups in much detail.