update identity mapping functions for EL and SL
The identity mapping functions copy the M and the AP bitfield to the
pesbt components and vice versa. Generally, this needs no changes when EL
and SL are added.

We should, however, make sure that EL and SL are stripped when
Zcherilevels is not used. It's safe to add such a check for both
directions.