update quadrant functions for risc-v cheri v0.9.0
Update the ap compression and decompression functions that use quadrant
encoding. Make the necessary changes for risc-v cheri v0.9.0. Process the new
LM permission and add the new encoding in quadrant 2.

Update the test cases for quadrant encoding and decoding.

This commit does not add SL and EL from the optional Zcherilevels extension.