update AP compression for EL, SL
Update the cc_N(m_ap_compress) function to handle the EL and SL
permissions that are used if Zcherilevels is supported.

Rewrite some of the checks for Q2 and Q3. The basic idea is to make the
checks for EL,SL=1 optional when we check for the quadrant and mandatory
when we check for an exact encoding.