remove unnecessary brackets from AP tests
We should use brackets consistently in the macros that define AP
compression and decompression tests. The test name (which must be unique)
is derived from the macro parameters. Unconsistent use of brackets makes
it harder to detect duplicate tests.

This patch changes the test definitions to use the minimum number of
brackets.