use define for shifting M in _m_ap
The test cases convert between separate values of m and ap and a joint
representation that we call _m_ap.

For 128r, _m_ap is simply M | AP where M is shiftted by the size of the
AP field. Use a define for AP's field size to make this clear.